Enemy Pathfinding Works Now
It's still laggy sometimes. I imagine it would be worse with more enemies, but it's a lot smoother than before. I'll try to change to A* and see if that works better. But I have a working pathfinding system now at least. The A* changes may take a while and won't be done in time for the next sprint.