Added a couple sound effects and fixed a bug
I added sounds for the player getting hurt, the enemies getting hit, and the bullets firing from the enemies. I couldn't get the walking sound to work, but the files I was trying to use are in there for that. Also, I fixed the bug wherein the starting slot in the hotbar starts on a different slot if you die and reset the game.