Start over button debugged
Debugged the start over button on my mac. All I did was recreate the scene and add the "Click" node from the menu scene that it was missing. If it's a lot fo work merging all you need to do to debug this is copy paste the "Click" sound from the menu scene and add it to the game over scene under the VBox containter named menu